More fixes on uml test. Seperate download files to a seperate script. Check the sha1 checksum after downloading. Using ./scriptname instead of sh scriptname. Don't redirect output to $LOG during setup.

+++ test/uml/demo.sh
@@ -5,25 +5,28 @@ set -x
 
 # One-time setup common to all tests
 . config_uml
-test -e  linux-${KERNEL_VERSION}/linux || sh build_uml.sh
-test -e uml_fs1 || { sh build_fs.sh uml_fs1; sudo sh build_fs_root.sh uml_fs1; }
+./download.sh
+test -e  linux-${KERNEL_VERSION}/linux || ./build_uml.sh
+test -e uml_fs1 || { ./build_fs.sh uml_fs1; sudo ./build_fs_root.sh uml_fs1; }
 
 # Just run each test once through, for a quick sanity check
 ITERATIONS=1
 export ITERATIONS
 
 # Single node tests
-sudo sh setup_network_root.sh 192.168.100.1 192.168.100.111 uml_1 uml_fs1
+sudo ./setup_network_root.sh 192.168.100.1 192.168.100.111 uml_1 uml_fs1
 . config_single
-sh test_agent_die.sh
+./test_agent_die.sh
+sleep 30
 
 # Two-node tests
-cp uml_fs1 uml_fs2
-chmod a+rw uml_fs2
-sudo sh setup_network_root.sh 192.168.100.2 192.168.100.111 uml_2 uml_fs2
+test -e uml_fs2 || { cp uml_fs1 uml_fs2; chmod a+rw uml_fs2; }
+sudo ./setup_network_root.sh 192.168.100.2 192.168.100.111 uml_2 uml_fs2
 . config_replication
-sh test_fullreplication.sh
-sh test_source_stop.sh
-sh test_target_stop.sh
+./test_fullreplication.sh
+sleep 30
+./test_source_stop.sh
+sleep 30
+./test_target_stop.sh
 
 echo All tests complete.
